sliderFix#2
